Move to web and do some cleanup
