+++ routes/web.php
@@ -8,7 +8,10 @@ use App\Http\Controllers\PostsController;
 use App\Http\Controllers\ProjectsController;
 use App\Http\Controllers\ProjectsPortfolioController;
 use App\Http\Controllers\RunCommandController;
+use Carbon\Carbon;
 use Illuminate\Support\Facades\Route;
+use Spatie\Sitemap\Sitemap;
+use Spatie\Sitemap\Tags\Url as SitemapUrl;
 
 /*
 |--------------------------------------------------------------------------
@@ -28,14 +31,14 @@ Route::view('/privacy', 'privacy')->name('privacy');
 
 Route::get('/projects', [ProjectsPortfolioController::class, 'index'])->name('portfolio');
 Route::prefix('/blog')
-	->group(function () {
+	->group(static function () {
 		Route::get('/', [BlogPostsController::class, 'index'])->name('blog');
 		Route::get('/{post:slug}', [BlogPostsController::class, 'show'])->name('blog.show');
 	});
 
 Route::prefix('/dashboard')
 	->middleware(['auth', 'verified'])
-	->group(function() {
+	->group(static function () {
 		Route::view('/', 'admin.dashboard')->name('dashboard');
 
 		Route::resource('commands', CommandController::class);
@@ -50,5 +53,31 @@ Route::prefix('/dashboard')
 
 		Route::resource('projects', ProjectsController::class);
 	});
+
+Route::get('/sitemap.xml',
+	static fn (Request $request) => Cache::remember('sitemap',
+		// cache for a week since the blog might update weekly
+		ttl()->days(7)->get(),
+
+		static fn (): Sitemap => Sitemap::create()
+
+			->add(SitemapUrl::create(route('home'))
+				->setLastModificationDate(Carbon::today())
+				->setChangeFrequency(SitemapUrl::CHANGE_FREQUENCY_WEEKLY)
+				->setPriority(0.1))
+
+			->add(SitemapUrl::create(route('privacy'))
+				->setLastModificationDate(Carbon::today())
+				->setChangeFrequency(SitemapUrl::CHANGE_FREQUENCY_MONTHLY)
+				->setPriority(0.1))
+
+			->add(SitemapUrl::create(route('blog'))
+				->setLastModificationDate(Carbon::today())
+				->setChangeFrequency(SitemapUrl::CHANGE_FREQUENCY_WEEKLY)
+				->setPriority(0.1))
+
+			->add(\App\Models\Post::all())
+	)
+		->toResponse($request));
 
 require __DIR__.'/auth.php';
